Adding Enter button functionality
When Enter is pressed it translate the current input and display it, and when Shift + Enter is pressed it goes to new line on input box and translate it. New line is add to output if we have new line in input box.

Added functionality to default translate from International to Morse.